+++ AddOns/LiteButtonAuras/CHANGELOG.md
@@ -1,6 +1,8 @@
 # LiteButtonAuras
 
-## [11.1.0-3](https://github.com/xod-wow/LiteButtonAuras/tree/11.1.0-3) (2025-04-01)
-[Full Changelog](https://github.com/xod-wow/LiteButtonAuras/compare/11.1.0-2...11.1.0-3) [Previous Releases](https://github.com/xod-wow/LiteButtonAuras/releases)
+## [11.1.5-1](https://github.com/xod-wow/LiteButtonAuras/tree/11.1.5-1) (2025-04-21)
+[Full Changelog](https://github.com/xod-wow/LiteButtonAuras/compare/11.1.0-3...11.1.5-1) [Previous Releases](https://github.com/xod-wow/LiteButtonAuras/releases)
 
-- Add wago.io uploader  
+- Fix UI options to refresh without reload, oops  
+- Update ToC for 11.1.5 and 1.15.7  
+- Use not UnitCanAttack instead of UnitIsFriend